### 2025-10-07 #9155 @kyleniemeyer
In the meantime, please take a look at the comments above ⬆️ from the editorialbot to address any DOI, license, or paper issues if you're able (there may not be any), or suggest reviewers.

For reviewers, please suggest 5 reviewers from the [database listed above](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network. Their GitHub handles are most useful to receive but **please** do not use "@" to reference them since it will prematurely ping them, and do not go asking folks to review yourself.

Unfortunately, we do not have any editors available to handle this right now, so I need to put this on the waitlist until someone becomes available.
