+++ paper/paper.bib
@@ -7,7 +7,8 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/141000671}
 }
 
 @article{lammps,
@@ -17,7 +18,8 @@
   volume={271},
   pages={108171},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cpc.2021.108171}
 }
 
 @article{gromacs1,
@@ -28,7 +30,8 @@
   number={16},
   pages={1701--1718},
   year={2005},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/jcc.20291}
 }
 
 @article{gromacs4,
@@ -38,7 +41,8 @@
   volume={1},
   pages={19--25},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.softx.2015.06.001}
 }
 
 @article{gromacs3,
@@ -49,7 +53,8 @@
   number={7},
   pages={845--854},
   year={2013},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btt055}
 }
 
 @article{gromacs2,
@@ -60,7 +65,8 @@
   number={8},
   pages={306--317},
   year={2001},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s008940100045}
 }
 
 @article{optim,
@@ -70,7 +76,8 @@
   volume={3},
   number={24},
   year={2018},
-  publisher={Open Journals}
+  publisher={Open Journals},
+  doi={10.21105/joss.00615}
 }
 
 @article{diffeq,
@@ -90,7 +97,8 @@
   volume={3},
   pages={1381--1384},
   year={1998},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/icassp.1998.681704}
 }
 
 @article{ase,
