### 2025-11-20 #9480 @editorialbot
[REVIEW]: YetAnotherSimulationSuite.jl: An Atomic Simulation Suite in Julia
**Submitting author:** @Cavenfish (<a href="http://orcid.org/0000-0002-7416-8629">Brian Ferrari</a>)
**Repository:** https://github.com/Cavenfish/YetAnotherSimulationSuite.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
