### 2025-11-20 Cavenfish/YetAnotherSimulationSuite.jl#3 @mikesha2
Novelty/advancement, benchmarking, force calculations
In reference to #openjournals/joss-reviews#9480,

First, you have clearly put in a lot of work into this repository in terms of taking care of tedious stuff like cell-based neighbor calculations and hard coding a few potentials. The documentation looks complete enough, and the organization is as JOSS requires.

My main concern is of course the most subjective part of the [evaluation](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort): does this work "either [enable] some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)." Why should I as a user choose your package over Molly.jl?

Questions I ask when choosing a software for molecular dynamics or quantum chemistry calculations:
**Hardware considerations**
1. How fast is the software? (unsure if you have any specific benchmarks you can point to, along with test system details)
2. What's the memory usage like? If I want to run modest-sized simulations locally, do I need to whip out my M4 Max with 128 GB RAM? What memory allocation would I need to assign on an HPC system? Or more specifically, what number do I put in my SLURM job script?
3. Is there GPU/FPGA acceleration?
**Day-to-day questions**
1. How easily can I add custom forces? Do I need to manually calculate the forces from the energy, or does autodiff take care of that (JAX-MD)?
2. How easily can I restrict degrees of freedom/apply constraints?
3. Are there built-in analysis pipelines I can use?
**Numerical stability**
1. How easily do NaNs pop up? Are forces/energies regularized so that my system doesn't blow up? Are NaNs propagated through the simulation, and how much of a performance hit is that?
**Style/usage, optional**
1. Is there some special way the code makes you think, some abstraction which more closely mirrors how a biophysicist thinks about their problem?

From looking at your paper and the code, I don't have easy answers to these questions.
