### 2025-11-20 Cavenfish/YetAnotherSimulationSuite.jl#3 @Cavenfish
Thank you for the recognition of the work I've put into this project! Also, thank you for agreeing to review this JOSS submission and doing so very quickly.

I'll first address your general question:
> Why should I as a user choose your package over Molly.jl?

With `Molly.jl` specifically, a user might choose YASS over it since it has the capability to perform geometry optimizations, cell optimizations, and harmonic frequency calculations. As far as I'm aware `Molly.jl` does not have those particular capabilities.

When comparing to other commonly used packages:
 - YASS is faster than pure ASE-based molecular dynamics
 - YASS is easier to compile and interact with than LAMMPS and GROMACs
 - YASS (through `Optim.jl`) provides more optimization algorithms for geometry/cell relaxations than JAX-MD

In some aspects YASS falls short compared with these other (older and more established) packages, but it still has something to offer.

In terms of your numbered questions:

 1. No benchmarks have been done yet. This can be done and added somewhere in the documentation or readme.
 2. Again, no explicit memory benchmarks have been done. In most of my slurm scripts I've used 2-4GB of memory depending on the potential, number of molecules, and simulation settings. If wanted, more precise measurements can be made and again added to the documentation or readme.
 3. Sadly, no.
 4. Very easily, an example is shown [here](https://cavenfish.github.io/YetAnotherSimulationSuite.jl/dev/potentials/customPot/) in the documentation. In the example the forces are calculated analytically, however, a user can easily use one of the [JuliaDiff](https://github.com/JuliaDiff) packages to perform autodiff-style force calculations. Currently YASS does not do this automatically for the user.
 5. This is a work in progress. Within the code there is the option to use fixed atom constraints, but nothing more as of now. YASS is setup to be flexible and easily customizable, so a user could in theory develop their own custom constraints but there is no documentation for this yet.
 6. Currently, YASS only has velocity autocorrelation function (VACF) analysis and radial distribution function (RDF) as a built-in analysis pipelines.
 7. The frequency of NaNs will heavily depend on the potential used. For the built-in potentials, they should (hopefully) never come up.  However, there isn't any checks for it in place currently nor any tests on the performance hit they cause.
 8. Tough for me to say. I feel the code is intuitive to work with, but I wrote it 😅. Honestly though, in this respect I don't think there is anything too special about YASS.

You mention:
>From looking at your paper and the code, I don't have easy answers to these questions.

Would you like me to try and include some of these answers into the paper? The word count limit might present some challenges but I can still try to add some into it.
