### 2025-11-24 f23a12f
add benchmarks and memory considerations to docs. relates to #3
