+++ benchmarks/yass/base.jl
@@ -1,0 +1,4 @@
+using Pkg
+Pkg.activate("./")
+
+using YetAnotherSimulationSuite
