+++ benchmarks/yass/Project.toml
@@ -1,0 +1,3 @@
+[deps]
+Optim = "429524aa-4258-5aef-a3af-852621145aeb"
+YetAnotherSimulationSuite = "d3ec4d54-14c5-4f6e-8d2e-82ae0926cfed"

[…]
