### 2025-11-24 af4b739
update statement of need to reflect YASS comparisons to other packages. relates to #3
