### 2025-11-24 Cavenfish/YetAnotherSimulationSuite.jl#3 @Cavenfish
I've made some edits to the documentation, readme, and paper  to address your comments.

Please let me know if things can still be improved!
