+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"bodies/#Simulation-Bodies","page":"Bodies","title":"Simulation Bodies","text":"","category":"section"},{"location":"bodies/","page":"Bodies","title":"Bodies","text":"YASS provides two main types of simulation objects: Particle and Cell. Each serves different purposes in molecular simulations:","category":"page"},{"location":"bodies/","page":"Bodies","title":"Bodies","text":"Particle: Represents individual atoms or particles\nCell: Represents periodic systems like crystals","category":"page"},{"location":"bodies/#Particle-Objects","page":"Bodies","title":"Particle Objects","text":"","category":"section"},{"location":"bodies/","page":"Bodies","title":"Bodies","text":"The Particle type is the fundamental building block for molecular simulations. Each particle has:","category":"page"},{"location":"bodies/","page":"Bodies","title":"Bodies","text":"Position vector (r)\nVelocity vector (v) \nMass (m)\nChemical symbol (s)","category":"page"},{"location":"bodies/#Creating-Particles","page":"Bodies","title":"Creating Particles","text":"","category":"section"},{"location":"bodies/","page":"Bodies","title":"Bodies","text":"You can create particles in several ways:","category":"page"},{"location":"bodies/","page":"Bodies","title":"Bodies","text":"using YetAnotherSimulationSuite\n\n# Read from XYZ file\natoms = readSystem(\"molecule.xyz\")\n\n# Create manually\nwater::Vector{MyAtoms} = [\n    Particle([0.000,  0.000, 0.000], zeros(3), 15.999, \"O\"),\n    Particle([0.757,  0.586, 0.000], zeros(3),  1.008, \"H\"),\n    Particle([0.757, -0.586, 0.000], zeros(3),  1.008, \"H\")\n]\n\n# Save to file\nwrite(\"water.xyz\", water)","category":"page"},{"location":"bodies/#Modifying-Particles","page":"Bodies","title":"Modifying Particles","text":"","category":"section"},{"location":"bodies/","page":"Bodies","title":"Bodies","text":"The Particle type is mutable, allowing modifications after creation:","category":"page"},{"location":"bodies/","page":"Bodies","title":"Bodies","text":"# Modify position\natoms[1].r .= [1.0, 0.0, 0.0]\n\n# Change velocity\natoms[1].v .= [0.1, 0.0, 0.0]\n\n# Update mass (e.g., for isotope studies)\natoms[1].m = 18.015  # Change to heavy water\n\n# Change chemical symbol\natoms[1].s = \"D\"     # Deuterium","category":"page"},{"location":"bodies/#Particle-Manipulation-Functions","page":"Bodies","title":"Particle Manipulation Functions","text":"","category":"section"},{"location":"bodies/","page":"Bodies","title":"Bodies","text":"YASS provides several utility functions for working with particles:","category":"page"},{"location":"bodies/","page":"Bodies","title":"Bodies","text":"# Translate all particles\ntranslate!(atoms, [1.0, 0.0, 0.0])\n\n# Center particles at origin\ncenterBdys!(atoms)\n\n# Swap positions of two particles\nswapAtoms!(atoms, 1, 2)\n\n# Change isotopes for specific atoms\nswapIso!(atoms, [1,2], [2.014, 2.014])  # Convert H to D\n\n# Get center of mass\ncom = CoM(atoms)\n\n# G
[…]
