### 2025-11-24 859085a
fix tables to bold rather than italics smaller time
