+++ docs/src/benchmark.md
@@ -23,9 +23,9 @@ Geometry optimizations using the LBFGS algorithm were performed on gold clusters
 
 | Number of Atoms | Iterations | YASS Time (s) | ASE Time (s) | YASS Speedup |
 |-----------------|------------|---------------|--------------| -------------|
-| 100             | 1500       | 11.8          | *8.4*        | 0.71x        |
-| 500             | 1500       | *29.6*        | 64.0         | 2.16x        |
-| 1000            | 1500       | *108.7*       | 218.8        | 2.01x        |
+| 100             | 1500       | 11.8          | **8.4**      | 0.71x        |
+| 500             | 1500       | **29.6**      | 64.0         | 2.16x        |
+| 1000            | 1500       | **108.7**     | 218.8        | 2.01x        |
 
 #### Harmonic Frequencies
 
@@ -33,9 +33,9 @@ Harmonic frequency calculations were performed on gold clusters containing 100, 
 
 | Number of Atoms | YASS Time (s) | ASE Time (s) | YASS Speedup |
 |-----------------|---------------|--------------| -------------|
-| 100             | 11.2          | *3.3*        | 0.29x        |
-| 500             | *22.7*        | 89.8         | 3.95x        |
-| 1000            | *87.1*        | 490.8        | 5.63x        |
+| 100             | 11.2          | **3.3**      | 0.29x        |
+| 500             | **22.7**      | 89.8         | 3.95x        |
+| 1000            | **87.1**      | 490.8        | 5.63x        |
 
 #### Molecular Dynamics
 
@@ -43,12 +43,12 @@ Molecular dynamics simulations were performed on gold clusters containing 100, 5
 
 | Number of Atoms | Time Steps | YASS Time (s) | ASE Time (s) | YASS Speedup |
 |-----------------|------------|---------------|--------------| -------------|
-| 100             | 1000       | 15.8          | *4.8*        | 0.30x        |
-| 100             | 2000       | 16.1          | *9.3*        | 0.58x        |
-| 100             | 5000       | *16.6*        | 22.4         | 1.35x        |
-| 500             | 1000       | *21.9*        | 32.3         | 1.47x        |
-| 500             | 2000       | *27.7*        | 65.9         | 2.38x        |
-| 500             | 5000       | *44.2*        | 160.7        | 3.63x        |
-| 1000            | 1000       | *35.6*        | 79.4         | 2.23x        |
-| 1000            | 2000       | *55.9*        | 165.3        | 2.95x        |
-| 1000            | 5000       | *112.5*       | 406.1        | 3.61x        |
+| 100             | 1000       | 15.8          | **4.8**      | 0.30x        |
+| 100             | 2000       | 16.1          | **9.3**      | 0.58x        |
+| 100             | 5000       | **16.6**      | 22.4         | 1.35x        |
+| 500             | 1000       | **21.9**      | 32.3         | 1.47x        |
+| 500             | 2000       | **27.7**      | 65.9         | 2.38x        |
+| 500             | 5000       | **44.2**      | 160.7        | 3.63x        |
+| 1000            | 1000       | **35.6**      | 79.4         | 2.23x        |
+| 1000            | 2000       | **55.9**      | 165.3        | 2.95x        |
+| 1000            | 5000       | **112.5**     | 406.1        | 3.61x        |
