### 2025-11-26 #9480 @andreww
@hisacro - please see a couple of the messages above - particularly https://github.com/openjournals/joss-reviews/issues/9480#issue-3647332658 which outlines the process and asks you to run `@editorialbot generate my checklist` as a comment in this issue to get started with the checklist driven review and https://github.com/openjournals/joss-reviews/issues/9480#issuecomment-3558132758 which is my introduction. We're looking to turn this around in two to three weeks and I'll check in here to see how things are going.
