### 2025-11-26 #9480 @jgreener64
Looks like an interesting project. Overall I agree with the points made in https://github.com/Cavenfish/YetAnotherSimulationSuite.jl/issues/3. Some comments:
- The package should be registered in the Julia general repo. Aside from anything else, the name will be discussed there and you don't want it to change after publication.
- The installation docs at https://cavenfish.github.io/YetAnotherSimulationSuite.jl/dev as written currently won't work until registration.
- The docs links in the readme text seem broken.
- Cite prior Julia simulation packages in the paper.
