### 2025-11-26 262ae7f
update readme installation
