### 2025-11-26 d9c5a67
add back in Molly and NQCDynamics citations
