### 2025-11-26 #9480 @Cavenfish
Hey @jgreener64, thank you for taking time to review my submission!

> The package should be registered in the Julia general repo. Aside from anything else, the name will be discussed there and you don't want it to change after publication.

The package is already registered in the general repo. The name was originally an issue, `YASS.jl` was renamed to the longer `YetAnotherSimulationSuite.jl`, so i truly appreciate you giving me a heads-up on that.

> The installation docs at https://cavenfish.github.io/YetAnotherSimulationSuite.jl/dev as written currently won't work until registration.

I can assure you they work. Can you try again? Maybe you need to run `update` to update the registry on your machine to match the general repo.

> The docs links in the readme text seem broken.

Thank you for spotting this! I've pushed a fix for it.

>  Cite prior Julia simulation packages in the paper.

Both `Molly.jl` and `NQCDynamics.jl` were originally cited, but on an edit I guess I forgot to put those citations back it. I've pushed an edit that puts them back in! If you know of more Julia simulation packages to cite please let me know!
