### 2025-11-27 0a69616
correct line in docs about YASS in general registry
