### 2025-11-27 7a72a6d
add link to dir with xyz files
