### 2025-12-08 b1bae49
bugfix: update getModePES and remove getModeInteractionPES
