### 2025-12-08 3c88f29
update docs to address reviewer comment
