### 2025-12-08 a4ee188
update examples to address reviewer comment
