### 2025-12-15 #9480 @andreww
@Cavenfish - some suggested changes for the paper text: https://github.com/Cavenfish/YetAnotherSimulationSuite.jl/pull/6

The PR also has a note about linking to 'dev' in the documentation. Not sure if you'll be able to (or want to) link to a version of the documentation tied to the release but something to think about.

I added some missing DOIs. Could you double check that remaining missing DOIs really don't exist (the editorial bot doesn't allways find them).
