### 2025-12-16 95c05cd
delete pre-release phrasing from docs and readme
