### 2025-12-16 6247823
build based on dd13579
