### 2025-12-16 f00b2a5
update docs links to stable
