+++ paper/paper.bib
@@ -129,6 +129,7 @@
   volume = "15",
   pages = "4897-4909",
   url = "https://doi.org/10.1039/D3SC05230C",
+  doi  = {10.1039/D3SC05230C},
 }
 
 @article{nqcdynamics,
