### 2025-12-16 #9480 @andreww
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: {JAX} {MD}: End-to-End Differentiable, Hardware Ac...
> ```

As far as I can tell, this one does not have a DOI
