### 2025-12-16 #9480 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
