### 2025-12-19 #9480 @kyleniemeyer
Hi @Cavenfish, there are a few things to address in the paper before publishing. In addition, the version number on the archive @ Zenodo (v1) does not match the version recorded here. If the archived version is v0.1.2, please update it in the metadata there, or else archive v0.1.2 and report the DOI here.

I made edits for the paper in https://github.com/Cavenfish/YetAnotherSimulationSuite.jl/pull/7, could you review and merge?

In addition, while I fixed the citation to the JAX MD paper, I question whether that should be cited, since [that paper was actually rejected](https://openreview.net/forum?id=r1xMnCNYvB)) (and the GitHub URL linked in it is broken). Instead, the [correct repo](https://github.com/jax-md/jax-md) references an arXiv preprint that can be cited instead.
