### 2025-12-19 3dae9ef
cite JAX MD preprint instead of rejected paper
