+++ paper/paper.bib
@@ -115,10 +115,14 @@
 }
 
 @misc{jaxmd,
-  title={{JAX} {MD}: End-to-End Differentiable, Hardware Accelerated, Molecular Dynamics in Pure {P}ython},
+  title={{JAX}, {M.D.}: A Framework for Differentiable Physics}, 
   author={Samuel S. Schoenholz and Ekin D. Cubuk},
   year={2020},
-  url={https://openreview.net/forum?id=r1xMnCNYvB}
+  eprint={1912.04232},
+  archivePrefix={arXiv},
+  primaryClass={physics.comp-ph},
+  url={https://arxiv.org/abs/1912.04232},
+  doi={10.48550/arXiv.1912.04232}
 }
 
 @article{molly,
