### 2025-12-19 #9480 @Cavenfish
Hey @kyleniemeyer,

I've merged the PR, and if I'm not mistaken I've corrected the metadata in Zenodo.

>In addition, while I fixed the citation to the JAX MD paper, I question whether that should be cited, since [that paper was actually rejected](https://openreview.net/forum?id=r1xMnCNYvB)) (and the GitHub URL linked in it is broken). Instead, the [correct repo](https://github.com/jax-md/jax-md) references an arXiv preprint that can be cited instead.

This was something I was unsure of when adding it. I've seen other papers cite the paper, maybe they cite the arXiv preprint you mention. I've changed the citation over to the preprint, hopefully I managed to get the title formatted properly.
